Fix renderer, add svg support
