+++ config/vite.renderer.config.mts
@@ -1,0 +1,31 @@
+import { defineConfig } from 'vite';
+import react from '@vitejs/plugin-react';
+import svgr from 'vite-plugin-svgr';
+import tsConfigPaths from 'vite-tsconfig-paths';
+import path from 'path';
+
+export default defineConfig({
+  root: path.resolve(__dirname, '../src/renderer'),
+  build: {
+    rollupOptions: {
+      input: {
+        index: path.resolve(__dirname, 'index.html'),
+      },
+      output: {
+        sourcemap: false,
+      },
+    },
+    commonjsOptions: {
+      include: [/node_modules/],
+      transformMixedEsModules: true,
+    },
+    sourcemap: false,
+  },
+  optimizeDeps: {
+    include: ['shallowequal', 'react-is', 'hoist-non-react-statics'],
+  },
+  resolve: {
+    extensions: ['.js', '.ts', '.jsx', '.tsx', '.css'],
+  },
+  plugins: [react(), svgr(), tsConfigPaths()],
+});
