+++ forge.config.ts
@@ -41,19 +41,19 @@ const config: ForgeConfig = {
         {
           // `entry` is just an alias for `build.lib.entry` in the corresponding file of `config`.
           entry: 'src/main/index.ts',
-          config: 'config/vite.main.config.ts',
+          config: 'config/vite.main.config.mts',
           target: 'main',
         },
         {
           entry: 'src/preload.ts',
-          config: 'config/vite.preload.config.ts',
+          config: 'config/vite.preload.config.mts',
           target: 'preload',
         },
       ],
       renderer: [
         {
           name: 'main_window',
-          config: 'config/vite.renderer.config.ts',
+          config: 'config/vite.renderer.config.mts',
         },
       ],
     }),
