+++ forge.env.d.ts
@@ -1,0 +1,1 @@
+/// <reference types="@electron-forge/plugin-vite/forge-vite-env" />
