+++ package.json
@@ -54,9 +54,9 @@
     "@electron-forge/maker-rpm": "^7.8.1",
     "@electron-forge/maker-squirrel": "^7.8.1",
     "@electron-forge/maker-zip": "^7.8.1",
-    "@electron-forge/plugin-vite": "^7.8.1",
     "@electron-forge/plugin-auto-unpack-natives": "^7.8.1",
     "@electron-forge/plugin-fuses": "^7.8.1",
+    "@electron-forge/plugin-vite": "^7.8.1",
     "@electron/fuses": "^1.8.0",
     "@types/lodash.groupby": "^4.6.9",
     "@types/lodash.uniq": "^4.5.9",
@@ -86,12 +86,14 @@
     "ts-node": "^10.9.2",
     "typescript": "^5.6.2",
     "url-loader": "^4.1.1",
-    "vite": "^5.4.19"
+    "vite": "^5.4.19",
+    "vite-plugin-svgr": "^4.3.0",
+    "vite-tsconfig-paths": "^5.1.4"
   },
   "dependencies": {
     "@electron-forge/publisher-github": "^7.8.1",
     "@hello-pangea/dnd": "^17.0.0",
-    "@pkmn/sim": "^0.9.15",
+    "@pkmn/sim": "^0.9.35",
     "@xyflow/react": "^12.3.2",
     "csv-parse": "^5.5.6",
     "csv-stringify": "^6.5.1",
@@ -102,8 +104,8 @@
     "i18next": "^23.15.1",
     "i18next-browser-languagedetector": "^8.0.0",
     "i18next-http-backend": "^2.6.2",
-    "react": "^18.3.1",
-    "react-dom": "^18.3.1",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-i18next": "^15.0.2",
     "react-is": "^18.3.1",
     "react-notifications-component": "^4.0.1",
