+++ src/assets.d.ts
@@ -1,11 +1,11 @@
-declare module '*.svg' {
+/*declare module '*.svg' {
   import React = require('react');
 
   export const content: SVGSVGElement;
   export const ReactComponent: React.FunctionComponent<React.SVGProps<SVGSVGElement>>;
   const src: string;
   export default src;
-}
+}*/
 
 declare module '*.png' {
   const content: unknown;
