+++ src/renderer/index.html
@@ -54,5 +54,6 @@
     <div id="tooltip"></div>
     <div id="context-menu"></div>
     <div id="tooltipContainer" popover="manual"></div>
+    <script type="module" src="main.tsx"></script>
   </body>
 </html>
